requirements.txt
pytorch
opencv-contrib-python
tqdm
numpy
pyautogui
mediapipe